Fix crash with appInfo() reading past end of buffer
This is a bug because the APDU proxy does not send back any flags unless
you use HAVE_BOLOS define when building the SDK for the device. This
does not seem to be built in on the NANO S PLUS target. That means using
appInfo() will crash because it assumes you get flags back but they are
never sent.

See supporting documentation and code here,
- https://github.com/LedgerHQ/ledger-secure-sdk/blob/cb5cfae5750d57e4c03460c5c93098fea7a8abcb/include/sdk_apdu_commands.h#L8-L32
- https://github.com/LedgerHQ/ledger-secure-sdk/blob/cb5cfae5750d57e4c03460c5c93098fea7a8abcb/src/os_io_seproxyhal.c#L1156-L1161